code/app-api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = ""
authors = ["Apparently Required <not-a-real-name@localhost>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
strawberry-graphql = {extras = ["debug-server"], version = "^0.229.1"}
uvicorn = {extras = ["standard"], version = "^0.27.1"}
couchbase = "^4.1.12"
openai = "^1.40.6"
cvxpy = "^1.5.2"
numpy = "^1.26.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.277"
pytest = "^7.4.0"
ipdb = "^0.13.13"
ipython = "^8.15.0"
pylint = "^3.2.6"

[tool.poetry.scripts]
app = 'app.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
